+++ addon/field-creator.js
@@ -483,7 +483,7 @@ class FieldOptionModal extends React.Component {
             })
           ),
           h("div", {className: "checkbox"},
-            h("label", null,
+            h("label", {className: "centerHorizontally"},
               h("input", {
                 type: "checkbox",
                 id: `${field.type.toLowerCase()}SortAlpha`,
@@ -495,7 +495,7 @@ class FieldOptionModal extends React.Component {
             )
           ),
           h("div", {className: "checkbox"},
-            h("label", null,
+            h("label", {className: "centerHorizontally"},
               h("input", {
                 type: "checkbox",
                 id: `${field.type.toLowerCase()}FirstValueDefault`,
@@ -557,7 +557,7 @@ class FieldOptionModal extends React.Component {
               type: "text",
               id: `${field.type.toLowerCase()}Length`,
               name: "length",
-              className: "form-control",
+              className: "form-control input-textBox",
               placeholder: "Max is 131,072 characters.",
               value: field.length,
               onChange: this.handleInputChange
@@ -569,7 +569,7 @@ class FieldOptionModal extends React.Component {
               type: "text",
               id: `${field.type.toLowerCase()}VisibleLines`,
               name: "vislines",
-              className: "form-control",
+              className: "form-control input-textBox",
               placeholder: "This field is required.",
               value: field.vislines,
               onChange: this.handleInputChange
@@ -630,7 +630,7 @@ class FieldOptionModal extends React.Component {
   renderRequiredCheckbox = () => {
     const {field} = this.state;
     return h("div", {className: "checkbox"},
-      h("label", null,
+      h("label", {className: "centerHorizontally"},
         h("input", {
           type: "checkbox",
           id: "required",
@@ -638,7 +638,7 @@ class FieldOptionModal extends React.Component {
           checked: field.required,
           onChange: this.handleInputChange
         }),
-        " Required"
+        "Required"
       )
     );
   };
@@ -646,7 +646,7 @@ class FieldOptionModal extends React.Component {
   renderUniqueCheckbox = () => {
     const {field} = this.state;
     return h("div", {className: "checkbox"},
-      h("label", null,
+      h("label", {className: "centerHorizontally"},
         h("input", {
           type: "checkbox",
           id: "unique",
@@ -654,7 +654,7 @@ class FieldOptionModal extends React.Component {
           checked: field.uniqueSetting,
           onChange: this.handleInputChange
         }),
-        " Unique"
+        "Unique"
       )
     );
   };
@@ -662,7 +662,7 @@ class FieldOptionModal extends React.Component {
   renderExternalIdCheckbox = () => {
     const {field} = this.state;
     return h("div", {className: "checkbox"},
-      h("label", null,
+      h("label", {className: "centerHorizontally"},
         h("input", {
           type: "checkbox",
           id: "externalId",
@@ -670,7 +670,7 @@ class FieldOptionModal extends React.Component {
           checked: field.external,
           onChange: this.handleInputChange
         }),
-        " External ID"
+        "External ID"
       )
     );
   };
@@ -688,13 +688,13 @@ class FieldOptionModal extends React.Component {
       className: "modal-dialog maxWidth500 maxHeight90vh overflowYAuto",
       onClick: (e) => e.stopPropagation()
     },
-    h("div", {className: "modal-content borderNone backgroundTransparent"},
-      h("div", {className: "modal-header borderBottomNone padding0_0_10_0 relativePosition"},
-        h("h4", {className: "modal-title textAlignCenter width100 margin0"}, "Set Field Options"),
+    h("div", {className: "modal-content relativePosition height100 flexColumn"},
+      h("div", {className: "modal-header flexSpaceBetween alignItemsCenter"},
+        h("h1", {className: "modal-title"}, "Set Field Options"),
         h("button", {
-          "aria-label": "Close Button",
           type: "button",
-          className: "close absoluteRightTop backgroundTransparent borderNone fontSize1_5 fontWeightBold cursorPointer",
+          "aria-label": "Close Set Field Options",
+          className: "close cursorPointer backgroundNone borderNone fontSize1_5 fontWeightBold",
           onClick: this.props.onClose
         },
         h("span", {"aria-hidden": "true"}, "×")